+++ lib/app/modules/main/bindings/main_binding.dart
@@ -1,7 +1,7 @@
 import 'package:duty_it/app/modules/calendar/controllers/calendar_view_controller.dart';
 import 'package:duty_it/app/modules/home/controllers/home_view_controller.dart';
 import 'package:duty_it/app/services/app_settings_service.dart';
-import 'package:duty_it/app/services/search_filter_service.dart';
+import 'package:duty_it/app/services/search_filter/search_filter_service.dart';
 import 'package:get/get.dart';
 
 import '../controllers/main_view_controller.dart';
