+++ lib/app/services/search_filter/models/search_filter.freezed.dart
